Pulls the BufferingScanTarget into a separate file.
